Prevent a bug due to unlinking a missing file
If we select GPL-3.0-or-later (for example), GPL-3.0-only would be
processed first, and the license/GPL3 file would be added to the
sources set for removal, because it was not excluded. Next,
GPL-3.0-or-later would be processed and would avoid adding the license
to sources going forward, but it would already be too late at that
point.

Here, we remove (or "discard") any existing matches in the sources set
if present, when we find the license that matches what was requested.